Adicionando o sistema de dano

+++ src/entidades/Entidade.java
@@ -38,10 +38,10 @@ public class Entidade {
 		}
 
 		public int getX() {
-			return this.x;
+			return (int)this.x;
 		}
 		public int getY() {
-			return this.y;
+			return (int)this.y;
 		}
 		public int getWidht() {
 			return this.widht;
